custom-elements/registering-custom-elements/definition-construction-algorithm-type-name-lowercased.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type should be converted to lower case</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Custom element type should be lowercased">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var GeneratedConstructor = doc.registerElement('X-A');
    var customElement = new GeneratedConstructor();

    assert_equals(customElement.localName, 'x-a', 'Custom element type should be lowercased');
}, 'Custom element type should be lowercased. Test element created via constructor');


test(function() {
    var doc = newHTMLDocument();
    var GeneratedConstructor = doc.registerElement('X-B');
    doc.body.innerHTML = '<X-B id="custom-element"></X-B>';
    var customElement = doc.querySelector('#custom-element');

    assert_equals(customElement.localName, 'x-b', 'Custom element type should be lowercased');
}, 'Custom element type should be lowercased. Test element created via innerHTML property');
</script>
</body>
</html>
